feat: add URL metadata tracking and 7-day auto-expiry

+++ src/main/java/arc/astra/jump/model/LeaderboardEntry.java
@@ -1,4 +1,4 @@
 package arc.astra.jump.model;
 
-public record LeaderboardEntry(String code, Double score) {
+public record LeaderboardEntry(String code, long score) {
 }
